+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Changed UI around with React and Material-UI
+- Asks to update first (can be disabled)
 
 ## [1.4.2] - 2021-04-02
 ### Changed